feat(weight-sync): give the colocated LoRA handler the extract/push split
LocalLoraWeightSync only exposed the fused sync(), so a colocated recipe had
to hold the trainer's parameters resident while the engine was already awake.
Splitting it lets the caller offload between the two halves, the same dance
unified_model already runs against RemoteLoraWeightSync.

The dispatch triple has to stay identical across both handlers or the dance
breaks silently, so it moves to the shared base and each handler keeps only
its own extract/push bodies.
